+++ R/43_use_vendor.R
@@ -21,8 +21,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' use_vendor("wurli/pedant")
-#' use_vendor("https://github.com/wurli/pedant", "R/add_double_colons.R")
+#' use_vendor("hadley/emo", "R/zzz.R")
 #' }
 use_vendor <- function(
   pkg,